status/keyboard: Fall back to deriving input source from system locale
If a user has configured their language outside of GNOME then their
input sources may not be set up.

But which input source to use, at least initially, is entirely derived
from which language the user has configured, so there's no reason we
can't infer the input source from the language.

This commit adds that inference as one final step before giving up
on figuring out the list of input sources.

This is useful in cases where e.g. GNOME Initial Setup gets called
on a live image that then gets installed leading to locale.conf
getting transferred over to the installed system, but not the live
user's gsettings.